@@ -11,47 +11,47 @@
   "author": "",
   "license": "ISC",
   "devDependencies": {
-    "@babel/core": "^7.4.5",
-    "@babel/plugin-transform-runtime": "^7.4.4",
-    "@babel/preset-env": "^7.4.5",
-    "@babel/preset-react": "^7.0.0",
-    "babel-loader": "^8.0.6",
-    "css-loader": "^0.28.11",
-    "html-loader": "^0.5.5",
-    "html-webpack-plugin": "^3.2.0",
-    "less": "^3.9.0",
-    "less-loader": "^4.0.5",
-    "style-loader": "^0.18.2",
-    "webpack": "^4.34.0",
-    "webpack-cli": "^3.3.4",
-    "webpack-dev-server": "^3.7.2"
+    "@babel/core": "7.18.6",
+    "@babel/plugin-transform-runtime": "^7.21.4",
+    "@babel/preset-env": "^7.21.4",
+    "@babel/preset-react": "^7.18.6",
+    "babel-loader": "^8.3.0",
+    "css-loader": "^6.7.3",
+    "html-loader": "^4.2.0",
+    "html-webpack-plugin": "^5.5.1",
+    "less": "^4.1.3",
+    "less-loader": "^11.1.0",
+    "style-loader": "^3.3.2",
+    "webpack": "^5.80.0",
+    "webpack-cli": "^5.0.2",
+    "webpack-dev-server": "^4.13.3"
   },
   "dependencies": {
-    "classnames": "^2.2.6",
+    "classnames": "^2.3.2",
+    "html-react-parser": "^3.0.16",
     "lodash.throttle": "^4.1.1",
     "moment": "^2.24.0",
-    "nav-datovelger": "^4.2.1",
-    "nav-frontend-core": "^4.0.9",
-    "nav-frontend-js-utils": "^1.0.7",
-    "nav-frontend-knapper": "^1.0.30",
-    "nav-frontend-knapper-style": "^0.3.26",
-    "nav-frontend-lenker-style": "^0.2.20",
-    "nav-frontend-paneler-style": "^0.3.15",
-    "nav-frontend-skjema": "^1.0.67",
-    "nav-frontend-skjema-style": "^1.0.40",
-    "nav-frontend-typografi": "^2.0.12",
-    "nav-frontend-typografi-style": "^1.0.13",
-    "nav-frontend-grid": "^1.0.20",
-    "nav-frontend-grid-style": "^0.2.18",
-    "nav-frontend-alertstriper": "^3.0.8",
-    "nav-frontend-alertstriper-style": "^2.0.17",
-    "nav-frontend-ikoner-assets": "^2.0.3",
+    "nav-datovelger": "^12.6.0",
+    "nav-frontend-core": "^6.0.1",
+    "nav-frontend-js-utils": "^1.0.20",
+    "nav-frontend-knapper": "^3.1.3",
+    "nav-frontend-knapper-style": "^2.1.2",
+    "nav-frontend-lenker-style": "^2.0.2",
+    "nav-frontend-paneler-style": "^2.0.2",
+    "nav-frontend-skjema": "^4.0.6",
+    "nav-frontend-skjema-style": "^3.0.3",
+    "nav-frontend-typografi": "^4.0.2",
+    "nav-frontend-typografi-style": "^2.0.2",
+    "nav-frontend-grid": "^2.0.2",
+    "nav-frontend-grid-style": "^1.0.3",
+    "nav-frontend-alertstriper": "^4.0.2",
+    "nav-frontend-alertstriper-style": "^3.0.2",
+    "nav-frontend-ikoner-assets": "^3.0.1",
     "pam-frontend-core": "^1.3.4",
-    "prop-types": "^15.7.2",
-    "react": "^16.8.6",
+    "prop-types": "^15.8.1",
+    "react": "^17.0.2",
     "react-day-picker": "^7.3.0",
-    "react-dom": "^16.8.6",
-    "react-router-dom": "^5.0.1",
-    "react-html-parser": "^2.0.2"
+    "react-dom": "^17.0.2",
+    "react-router-dom": "^6.10.0"
   }
 }
