@@ -1,6 +1,6 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import { BrowserRouter, Route, Switch } from 'react-router-dom';
+import { BrowserRouter, Route, Routes } from 'react-router-dom';
 import Preview from './preview/Preview';
 import { StoreProvider } from './context/StoreContext';
 import './styles.less';
@@ -9,11 +9,10 @@ import './styles.less';
 const Index = () => {
     return (
         <main>
-            <Switch>
-                <Route exact path={`${__CONTEXT_PATH__}/:uuid`} component={Preview} />
-            </Switch>
+            <Routes>
+                <Route exact path={`${__CONTEXT_PATH__}/:uuid`} element={<Preview />}/>
+            </Routes>
         </main>
-
     );
 };
 
